i18Next Commit Hook
Add separate configs for English and other languages: English should
populate default values, but other languages should not.

Scan in a pre-commit hook to make sure languages are set up correctly.

Language config file changes in this commit are evidence that the commit
hook is working -- the changes are due to new config we have to sort
keys.
