+++ .husky/pre-commit
@@ -2,3 +2,6 @@
 . "$(dirname -- "$0")/_/husky.sh"
 
 npx lint-staged
+
+yarn dlx i18next-scanner --config i18next-scanner.config-en.js
+yarn dlx i18next-scanner --config i18next-scanner.config-langs.js
